% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TRAFIC.r
\name{trafic}
\alias{trafic}
\title{TRAFIC: Test for Rare-variant Asoociation with Family Internal Control.}
\usage{
trafic(ped_file = "./inst/data_sibpair.ped",
  label_file = "./inst/data_sibpair.dat", ibd_file = "./inst/S_sibpair.ibd",
  gene_group = "./inst/data_sibpair.gene", impute = FALSE)
}
\arguments{
\item{ped_file}{a ped file contains genotypes.}

\item{label_file}{a file contains the name for the SNPs.}

\item{ibd_file}{a file contains the IBD status of each sibpair for every gene}

\item{gene_group}{a file contains the snps of interest in each gene}

\item{impute}{whether to generate imputed dosage}
}
\value{
\code{trafic} returns the allele frequecy for cases and controls as well as p-value for each tested gene
}
\description{
\code{trafic} Use TRAFIC to test for associations and returns the p-value.
}
\details{
\code{trafic} uses multiple impulation to impute sharing status of ambiguous
double-heterozygote who share one IBD chromosome region.
}
\examples{
#read in files
ped_file <- system.file("data_sibpair.ped", package="TRAFIC")
label_file <- system.file("data_sibpair.dat", package="TRAFIC")
ibd_file <- system.file("S_sibpair.ibd", package="TRAFIC")
gene_group <- system.file("data_sibpair.gene", package="TRAFIC")

#perform trafic
trafic(ped_file, label_file, ibd_file, gene_group, impute=FALSE)
}

