% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TRAFIC.r
\name{TRAFIC}
\alias{TRAFIC}
\title{Test for Rare-variant Asoociation with Family Internal Control.}
\usage{
TRAFIC(genotype_file = "data_sibpair.geno", label_file = "data_sibpair.ped",
  ibd_file = "S_sibpair.ibd", snp = NA)
}
\arguments{
\item{label_file}{filename for the SNPs.}

\item{genotype_filename}{filename for the genotype.}

\item{filename}{for the IBD status}

\item{names}{for the snps of interest}
}
\value{
TRAFIC return the p-value \link{http://test.com}.
}
\description{
\code{TRAFIC} returns the p-value.
}
\details{
TRAFIC uses multiple impulation to impute sharing status of ambiguous
double-heterozygote who share one IBD chromosome region.
}
\examples{
TRAFIC(genotype_file="genotype_test.geno", label_file="genotype_test.dat", ibd_file="S_sibpair.ibd", snp="1")
}

